Properly handles closing program
No longer keeps zombie process for http server

+++ testEnvSetup.py
@@ -6,10 +6,11 @@ import argparse
 import io
 import qrcode
 import socket
+import signal
+import sys
 from dotenv import load_dotenv
 
 load_dotenv()
-
 
 def start_ngrok(port, ngrok_path):
     # Start ngrok process
@@ -66,8 +67,6 @@ def start_ffmpeg_server(directory, ffmpeg_command):
         print(f"Error: {e}")
         return None
 
-
-
 def stop_ffmpeg(ffmpeg_server):
     try:
         # Send termination signal
@@ -107,9 +106,20 @@ def printURLQrCode(url):
     f.seek(0)
     print(f.read())
 
+# Signal handler for graceful shutdown
+def signal_handler(sig, frame):
+    global http_server, ffmpeg_server, ngrok, skip_ngrok
+    if not skip_ngrok:
+        stop_ngrok(ngrok)
+    if http_server:
+        http_server.terminate()
+        http_server.wait()  # Ensure it's fully terminated
+    if ffmpeg_server:
+        stop_ffmpeg(ffmpeg_server)
+    print("Ngrok, HTTP server, and FFmpeg server have been stopped.")
+    sys.exit(0)
 
 if __name__ == "__main__":
-
     parser = argparse.ArgumentParser(description="Start video streaming with ngrok, HTTP server, and FFmpeg.")
     parser.add_argument("--skip_ngrok", action='store_true', help="Skip running ngrok if specified.")
     args = parser.parse_args()
@@ -119,6 +129,13 @@ if __name__ == "__main__":
     ngrok_path = os.getenv('NGROK_PATH') 
     webcam_name = os.getenv('WEBCAM_NAME')
     ffmpeg_command = os.getenv('FFMPEG_COMMAND')
+
+    # Register signal handler
+    signal.signal(signal.SIGINT, signal_handler)
+
+    http_server = None
+    ffmpeg_server = None
+    ngrok = None
 
     if not skip_ngrok:
         ngrok, ngrok_url = start_ngrok(port, ngrok_path)
@@ -141,15 +158,15 @@ if __name__ == "__main__":
     http_server = start_http_server(video_streaming_dir, port)
 
     ffmpeg_server = start_ffmpeg_server(f'./videoStreaming/ffmpeg', ffmpeg_command)
-        
+
     input("Press Enter to stop ngrok...")
-    
+
     if not skip_ngrok:
         stop_ngrok(ngrok)
-    
-    http_server.terminate()
-    
-    stop_ffmpeg(ffmpeg_server)
+    if http_server:
+        http_server.terminate()
+        http_server.wait()
+    if ffmpeg_server:
+        stop_ffmpeg(ffmpeg_server)
 
     print("Ngrok, HTTP server, and FFmpeg server have been stopped.")
-
